refactor(hooks): split bash.rs into per-category module
